+++ book/src/applet/prelude/timer.rs
@@ -28,7 +28,7 @@ use alloc::rc::Rc;
 use core::cell::Cell;
 use core::time::Duration;
 
-fn main() {
+fn main() -> Result<(), Error> {
     assert!(button::count() > 0, "Board has no buttons.");
     assert!(led::count() > 0, "Board has no LEDs.");
 
@@ -66,8 +66,7 @@ fn main() {
                 button::Released if pressed.get() => released.set(true),
                 button::Released => (),
             }
-        })
-        .unwrap();
+        })?;
         //} ANCHOR_END: button
 
         //{ ANCHOR: pressed