+++ examples/rust/led/src/lib.rs
@@ -30,7 +30,7 @@ wasefire::applet!();
 use alloc::boxed::Box;
 use core::cell::Cell;
 
-fn main() {
+fn main() -> Result<(), Error> {
     // Make sure there is at least one button.
     let num_buttons = button::count();
     assert!(num_buttons > 0, "Board has no buttons.");
@@ -57,6 +57,8 @@ fn main() {
         };
 
         // We indefinitely listen by creating and leaking a listener.
-        button::Listener::new(button_index, handler).unwrap().leak();
+        button::Listener::new(button_index, handler)?.leak();
     }
+
+    Ok(())
 }